+++ travelhub_back/src/main/java/com/travelhub/repository/UserRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.travelhub.entity.User;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUserName(String userName);
-    Optional<User> findByUserId(Long userId);
+    Optional<User> findByUserId(Integer userId);
     Optional<User> findByUserEmail(String userEmail);
 }
